requirements.txt
requests==2.31.0
yfinance==0.2.38
pandas==2.1.4
matplotlib==3.8.2
mplfinance==0.12.10b0
pytrends==4.9.2
beautifulsoup4==4.12.3
lxml==5.1.0
schedule==1.2.1
python-dotenv==1.0.0
Pillow==10.2.0
cloudinary==1.38.0
twstock==1.3.3
APScheduler==3.10.4
